add eslint-plugin-jest and jest rules
