+++ .eslintrc.js
@@ -31,6 +31,8 @@ module.exports = {
         'plugin:@typescript-eslint/recommended',
         // Handle prettier rules through eslint https://github.com/prettier/eslint-plugin-prettier/blob/master/eslint-plugin-prettier.js#L65
         'plugin:prettier/recommended',
+        // add special jest rules https://github.com/jest-community/eslint-plugin-jest#rules
+        'plugin:jest/recommended',
       ],
       rules: {
         '@typescript-eslint/explicit-function-return-type': 'off',
@@ -50,6 +52,13 @@ module.exports = {
         '@typescript-eslint/require-await': 'error',
         '@typescript-eslint/no-empty-function': 'off',
         '@typescript-eslint/ban-types': 'off', // TODO: turn on
+        /** jest */
+        'jest/valid-title': 'off', // allow function names to be used as test titles
+        'jest/no-conditional-expect': 'off', // best practice, but TODO
+        'jest/no-alias-methods': 'off', // best practice, but TODO
+        'jest/expect-expect': 'off', // sometimes we compose assertion functions
+        'jest/no-disabled-tests': process.env.CI ? 'error' : 'off', // fail CI if tests are disabled, but do not interfere with local development
+        'jest/no-focused-tests': process.env.CI ? 'error' : 'off',
       },
       overrides: [
         {
