+++ package.json
@@ -48,6 +48,7 @@
     "eslint": "^8.14.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.2",
     "eslint-plugin-prettier": "^4.0.0",
     "express": "^4.18.2",
     "get-monorepo-packages": "^1.2.0",
