+++ packages/browser/src/lib/client-hints/__tests__/index.test.ts
@@ -1,29 +1,9 @@
 import { clientHints } from '..'
-import { UADataValues, UALowEntropyJSON } from '../interfaces'
-
-export const lowEntropyTestData: UALowEntropyJSON = {
-  brands: [
-    {
-      brand: 'Google Chrome',
-      version: '113',
-    },
-    {
-      brand: 'Chromium',
-      version: '113',
-    },
-    {
-      brand: 'Not-A.Brand',
-      version: '24',
-    },
-  ],
-  mobile: false,
-  platform: 'macOS',
-}
-
-export const highEntropyTestData: UADataValues = {
-  architecture: 'x86',
-  bitness: '64',
-}
+import {
+  highEntropyTestData,
+  lowEntropyTestData,
+} from '../../../test-helpers/client-hints'
+import { UADataValues } from '../interfaces'
 
 describe('Client Hints API', () => {
   beforeEach(() => {
